cleanup and reorganize types
