+++ types/src/legacy_types/account.rs
@@ -171,25 +171,6 @@ impl ValConfigs {
         return Ok(configs);
     }
 
-    // /// create the encrypted network address for use on the validator network.
-    // pub fn make_val_network_addr(ip_address: &Ipv4Addr, val_pubkey: PublicKey) -> NetworkAddress {
-    //           // Create the list of validator addresses
-    //     let val_network_string = format!("/ip4/{}/tcp/{}", ip_address.to_string(), DEFAULT_VAL_PORT);
-    //     let val_addr_obj: NetworkAddress = val_network_string
-    //         .parse()
-    //         .expect("could not parse validator network address");
-    //     val_addr_obj.append_prod_protos(val_pubkey, 0)
-    // }
-
-    // /// format the fullnode address which the validator's VFN will use.
-    // pub fn make_fullnode_unencrypted_addr(ip_address: &Ipv4Addr, fn_pubkey: PublicKey) -> NetworkAddress {
-    //     let fullnode_network_string = format!("/ip4/{}/tcp/{}", ip_address.to_string(), DEFAULT_VFN_PORT);
-    //     let fn_addr_obj: NetworkAddress = fullnode_network_string
-    //         .parse()
-    //         .expect("could not parse fullnode network address");
-    //     fn_addr_obj.append_prod_protos(fn_pubkey, 0)
-    // }
-
     /// format the fullnode address which the validator's VFN will use.
     pub fn make_unencrypted_addr(
         ip_address: &Ipv4Addr,
