+++ types/src/legacy_types/mod.rs
@@ -1,26 +1,5 @@
 //! `legacy-types`
 
-pub mod ancestry_legacy;
-pub mod app_cfg;
 pub mod block;
-pub mod burn;
-pub mod cumulative_deposits;
-pub mod donor_voice;
-pub mod donor_voice_txs;
-pub mod fee_maker;
-pub mod fixtures;
-pub mod jail;
-pub mod legacy_currency_info;
-pub mod legacy_recovery_v6;
-pub mod match_index;
-pub mod mode_ol;
-pub mod network_playlist;
-pub mod ol_account;
-pub mod pledge_account;
-pub mod proof_of_fee;
-pub mod receipts;
 pub mod tx_error;
 pub mod validator_config;
-pub mod validator_universe;
-pub mod vouch;
-pub mod wallet;
